gedit 3.4.2, fix builtin resources
This solves the problem with the missing interface in 3.4.[01].  The
problem was that the static GResources, which were to be initialized
via __attribute__((constructor)), are in libgedit-private.  The latter
didn't have -no-undefined, so ended up static, so when it was linked into
gedit.exe, the ctor/dtor symbols were dropped (similar to Vinagre 3.2 in
BGO#653558). If libgedit-private is shared, the ctor/dtor are exported.

So instead of exporting everything from gedit.exe, we can link against
a shared libgedit-private, allowing us to drop most of the implib hackery
previously required.